pyproject.toml
[project]
name = "api-user-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic==1.12.1",
    "anyio>=4.9.0",
    "asyncpg==0.29.0",
    "bcrypt==4.0.1",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.7",
    "pydantic[email]==2.5.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "sqlalchemy[asyncio]==2.0.23",
    "twisted>=24.11.0",
    "uvicorn>=0.34.2",
]
